requirements.txt
asgiref==3.7.2
bcrypt==4.2.0
beautifulsoup4==4.12.2
boto3==1.34.49
botocore==1.34.49
cffi==1.17.0
crispy-bootstrap5==2023.10
cryptography==43.0.0
deprecation==2.1.0
dj-database-url==2.1.0
Django==4.2.15
django-bootstrap-v5==1.0.11
django-crispy-forms==2.1
django-email-verification==0.3.3
django-formtools==2.5.1
django-storages==1.14.2
jmespath==1.0.1
packaging==23.2
paramiko==3.4.1
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.8.2
s3transfer==0.10.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==1.26.18
validators==0.22.0
